Add abstract analyser test base class
